% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centreScaleDiagrams.R
\name{centre_scale_diagrams}
\alias{centre_scale_diagrams}
\title{Centre and Scale the Extended Persistence Diagrams of an image.}
\usage{
centre_scale_diagrams(diagrams, scale = TRUE, scale_constant = 1)
}
\arguments{
\item{diagrams}{A list of extended persistence diagrams on a single image
computed in N evenly spaced directions.}

\item{scale_constant}{A constant to control the scaling of the diagram.
Default is 1. Note this is not the percentage of the original scale of the
image.}
}
\value{
A list of extended persistence diagrams of the same image, centred
at the origin and scaled by an amount proportional to the sum of the
minimum birth times.
}
\description{
Given a collection of persistence diagrams in N evenly spaced directions,
shift the points in each diagram so that the correspond to the points of the
persistence diagrams of the centred image. Then scale the birth and death
times of each point by some factor.
}
\details{
Centring and scaling the persistence diagrams allows for a more accurate
comparison of their shapes.
}
