% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceMatrix.R
\name{computeDistanceMatrix}
\alias{computeDistanceMatrix}
\title{Compute the q-Wasserstein Distance between the extended persistent homology
transform of multiple objects.}
\usage{
computeDistanceMatrix(diagrams, nObjects, q = 1, aligned = FALSE)
}
\arguments{
\item{diagrams}{\code{list} of extended persistence diagrams of multiple objects.
Each object must have their extended persistence diagrams computed in the
same number of directions. If \eqn{K} directions were used, then the first
\eqn{K} entries must correspond to the first object, the next \eqn{K}
entries to the next object and so on.}

\item{nObjects}{\code{int} giving the total number of objects. Used to determine
the number of directions.}

\item{q}{\code{numeric} determining the \eqn{L^q} distance to use. Must have
\eqn{1\leq q < \infty}. The default value is 1.}

\item{aligned}{\code{bool} to indicate whether the original images have the same
alignment. The default value is \code{FALSE}.}
}
\description{
Given a collection of objects \eqn{O_1,\dots,O_n} and directions
\eqn{v_1,\dots,v_K}, compute the distance matrix \eqn{D}$ between all pairs
of objects \eqn{O_i,O_j} for \eqn{i\neq j}. The objects \eqn{O_i} and
\eqn{O_j} have extended persistence diagrams \eqn{X_{i1},\dots,X_{iK}} and
\eqn{X_{j1},\dots,X_{jK}}, respectively. The distance between $O_i$ and
$O_j$ is the average distance between their diagrams.
}
