% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendedPersistence.R
\name{computeExtendedPersistence}
\alias{computeExtendedPersistence}
\title{Compute the Extended Persistence Diagrams of a given image}
\usage{
computeExtendedPersistence(
  bdry_curves,
  img_name,
  n_dirs,
  tolerance = 1/sqrt(2),
  save_output = FALSE,
  output_dir = NULL,
  f_name = NULL,
  verbose = TRUE
)
}
\arguments{
\item{bdry_curves}{A list containing points along the boundary curves of the
image. The list must have length number_of_curves + 1 with the first entry a
vector whose entries are the numbers of positively oriented curves and
negatively oriented curves, respectively. All of the positively oriented
curves must be listed before the negatively oriented curves.}

\item{img_name}{A string giving the name of the image.}

\item{n_dirs}{An integer giving the number of directions to use. This
must be positive and even.}

\item{tolerance}{Parameter to control the noise from the image. Any vertices
closer than this in a particular direction will be classified as noise and
ignored. (Default: 1/sqrt(2))}

\item{save_output}{If TRUE, will save output to directory specified by
output_dir. (Default: FALSE)}

\item{output_dir}{The directory to save the output. If save_output is TRUE and
no directory is specified, saves to working directory. (Default: NULL)}

\item{f_name}{The name of the output file saved. If save_output is TRUE and no
filename specified, prompts user for filename. (Default: NULL)}

\item{verbose}{If TRUE, prints indictors of progress throughout.
(Default: TRUE)}
}
\value{
A list of length 5 with names "Name", "Ord0", "Rel1", "Ext0",
"Ext1". Each entry is a list
containing the relevant part of the extended persistence diagram. The i-th
entry corresponds to the i-th direction.
}
\description{
Given a disjoint collection of simple closed curves representing the
boundaries of a two dimensional shape, compute the zeroth order extended
persistent homology transform of the shape in a given number of directions.
}
