% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rankFunctions.R
\name{discretisedRank}
\alias{discretisedRank}
\title{discretisedRank}
\usage{
discretisedRank(diagram, xcoords, ycoords)
}
\arguments{
\item{diagram}{A single persistence diagram with n points, stored as an n x 2 array.}

\item{xcoords}{A list of x-coordinates}

\item{ycoords}{A list of y-coordinates}
}
\value{
An array containing the values of the rank function, where the (i,j)-entry is the value at the point (x_i, y_j).
}
\description{
Evaluate the a discretised rank function, evaluated on the cartesian product of xcoord and ycoord.
}
