% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractBoundry.R
\name{extractBoundary}
\alias{extractBoundary}
\title{Extract the boundary curves of a binary image}
\usage{
extractBoundary(
  img,
  background = 1,
  saveOutput = FALSE,
  outputDir = NULL,
  fName = NULL,
  verbose = TRUE
)
}
\arguments{
\item{img}{A binary image of class `cimg' read in using
\code{\link[imager:load.image]{imager::load.image()}}.}

\item{background}{The value of the background, 0 for black and 1 for white.
The default value is \code{1}.}

\item{saveOutput}{If TRUE, will save output as a \code{.RDS} file to
directory specified by outputDir. The default values is \code{FALSE}.}

\item{outputDir}{The directory to save the output. If saveOutput is TRUE
and no directory is specified, saves to working directory. The default
is \code{NULL}.}

\item{fName}{The name of the output file saved. If saveOutput is TRUE and
no filename specified, prompts user for filename. The default value is
\code{NULL}.}

\item{verbose}{If TRUE, prints indictors of progress throughout. The default
value is \code{TRUE}.}
}
\value{
A list containing the points on the boundary curves structured as:
\itemize{
\item The first entry is a vector \code{(m,n)} containing the number of
anticlockwise (m) and clockwise (n) curves.
\item Entries \code{2} to \code{m+1}  contain the anticlockwise curves.
\item Entries \code{m+2} to \code{n+1} contain the clockwise curves.
}
}
\description{
Given a binary image (pixel values 0 and 1) \code{extractBoundary}
will trace all boundary curves along the midpoints of pixel edges between
foreground and background regions. Curves which bound the outside of a
foreground region are traced in the anticlockwise direction while curves
which bound the inside of a foreground region are traced in the clockwise
direction. The image is padded to ensure outermost pixels are part of the
background.
}
\seealso{
\code{\link[=multiExtractBoundary]{multiExtractBoundary()}}, \code{\link[imager:load.image]{imager::load.image()}}
}
