% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.extendedPHT.R
\name{getDefaultColours}
\alias{getDefaultColours}
\title{Get the Default Colours used in Plotting Extended Persistence Diagrams}
\usage{
getDefaultColours()
}
\value{
Vector containing the default colours used for plotting extended
persistence diagrams in the order (Ord0, Rel1, Ess0, Ess1).
}
\description{
Get the Default Colours used in Plotting Extended Persistence Diagrams
}
