% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractBoundry.R
\name{multiExtractBoundary}
\alias{multiExtractBoundary}
\title{Extract the boundary curves of binary images in a given directory}
\usage{
multiExtractBoundary(
  inputDir,
  imgType = "png",
  background = 0,
  saveOutput = FALSE,
  outputDir = NULL,
  verbose = TRUE
)
}
\arguments{
\item{inputDir}{The directory containing the images.}

\item{imgType}{The image file type for input files, specified without a
".". The default value is \code{png}.}

\item{background}{The value of the background, 0 for black and 1 for white.
The default value is \code{1}.}

\item{saveOutput}{If TRUE, will save output as a \code{.RDS} file to
directory specified by outputDir. The default values is \code{FALSE}.}

\item{outputDir}{The directory to save the output. If saveOutput is TRUE
and no directory is specified, saves to working directory. The default
is \code{NULL}.}

\item{verbose}{If TRUE, prints indictors of progress throughout. The default
value is \code{TRUE}.}
}
\value{
A list for each image containing the points of the boundary curves.
Stored in individual files (if saveOutput = TRUE) or a list.
}
\description{
Runs \code{\link[=extractBoundary]{extractBoundary()}} on all images of a specified type in a
given directory. Images must be binary (pixel values 0 and 1). For more
information, see \code{\link[=extractBoundary]{extractBoundary()}}
}
