% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.extendedDiagram.R
\name{plot.extendedDiagram}
\alias{plot.extendedDiagram}
\title{Plot an Extended Persistence Diagram or Barcode}
\usage{
\method{plot}{extendedDiagram}(
  x,
  barcode = FALSE,
  col = c("#000000", "#e69f00", "#009e73", "#d55e00"),
  pch = 15:18,
  lab.line = 2.2,
  ...
)
}
\arguments{
\item{x}{\code{extendedDiagram} object to plot.}

\item{barcode}{If \code{TRUE}, then barcode is plotted. Default value is \code{FALSE}.}

\item{col}{\code{vector} containing the four colours to use in plotting in the
order Ord0, Rel1, Ext0, Ext1. The default value is
\itemize{
\item "#000000" for Ord0
\item "#e69f00" for Rel1
\item "#009e73" for Ext0
\item "#d55e00" for Ext1
}}

\item{pch}{Symbols to use if plotting extended persistence diagram in order
Ord0, Rel1, Ext0, Ext1. The default value is \code{15:18}.}

\item{lab.line}{Control distance of axis labels to axes. The default value is
2.2.}

\item{...}{other \link[graphics]{graphical parameters} (see \code{\link[graphics]{par}} and
    section \sQuote{Details} below).}
}
\description{
The function \code{\link[=plot.extendedDiagram]{plot.extendedDiagram()}} plots the extended persistence diagram
or extended persistence barcode of an \code{extendedDiagram} object.
}
\details{
This is a generic function. Plot the the extended persistence diagram or the
extended persistence barcode of an object. If the diagram is plotted, then
the diagonal \eqn{birth = death} is also plotted. The classes in Ord0 and
Ext0 have \eqn{birth < death}  and are plotted above the diagonal, whereas
classes in Rel1 and Ext1 have \eqn{birth > death} and are plotted below the
diagonal. If the barcode is plotted then classes will be stacked in the
order Ord0, Rel1, Ext0, Ext1.
}
