% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distanceMatrix.R
\name{stackDiagrams}
\alias{stackDiagrams}
\title{Stack Multiple XPHTs into a Single List}
\usage{
stackDiagrams(inputDir)
}
\arguments{
\item{inputDir}{The directory containing the XPHTs stored as \code{.RDS} files.}
}
\value{
A list of XPHTs
}
\description{
Given a directory of XPHTs, load each XPHT and store as an element in a
list.
}
