% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centreScaleDiagrams.R
\name{centreScaleDiagrams}
\alias{centreScaleDiagrams}
\title{Centre and Scale Extended Persistence Diagrams}
\usage{
centreScaleDiagrams(
  diagrams,
  scale = TRUE,
  scaleConstant = 1,
  saveOutput = FALSE,
  outputDir = NULL,
  fName = NULL,
  verbose = TRUE
)
}
\arguments{
\item{diagrams}{A list of extended persistence diagrams on a single image
computed in \eqn{N} evenly spaced directions.}

\item{scale}{Flag to indicate whether diagrams are to be scaled or not. The
default value is \code{TRUE}.}

\item{scaleConstant}{\code{numeric} constant to control the scaling of the
diagram. This is \emph{not} the percentage of the original scale of the image.
The default value is 1.}

\item{saveOutput}{If TRUE, will save output as a \code{.RDS} file to
directory specified by outputDir. The default values is \code{FALSE}.}

\item{outputDir}{The directory to save the output. If saveOutput is TRUE
and no directory is specified, saves to working directory. The default
is \code{NULL}.}

\item{fName}{The name of the output file saved. If saveOutput is TRUE and
no filename specified, prompts user for filename. The default value is
\code{NULL}.}

\item{verbose}{If TRUE, prints indicators of progress throughout. The default
value is \code{TRUE}.}
}
\value{
A list of extended persistence diagrams centred at the origin and
with birth and death times scaled by an amount proportional to the sum of the
minimum birth times.
}
\description{
The function \code{\link[=centreScaleDiagrams]{centreScaleDiagrams()}} translates the points in an extended
persistence diagram to be centred about the origin and scales the points
to make different diagrams comparable.
}
\details{
Given a collection of persistence diagrams in \eqn{N} evenly spaced
directions, shift the points in each diagram so that they correspond to the
points of the persistence diagrams of the centred image. This is equivalent
to centring the convex hull of the original image. Then scale the
birth and death times of each point by some factor.
}
