% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centreScaleDiagrams.R
\name{centreScaleDiagrams}
\alias{centreScaleDiagrams}
\title{Centre and Scale the Extended Persistence Diagrams of an image.}
\usage{
centreScaleDiagrams(diagrams, scale = TRUE, scaleConstant = 1)
}
\arguments{
\item{diagrams}{A list of extended persistence diagrams on a single image
computed in \eqn{N} evenly spaced directions.}

\item{scale}{Flag to indicate whether diagrams are to be scaled or not. The
default value is \code{TRUE}.}

\item{scaleConstant}{\code{numeric} constant to control the scaling of the
diagram. This is \emph{not} the percentage of the original scale of the image.
The default value is 1.}
}
\value{
A list of extended persistence diagrams centred at the origin and
with birth and death times scaled by an amount proportional to the sum of the
minimum birth times.
}
\description{
Given a collection of persistence diagrams in \eqn{N} evenly spaced
directions, shift the points in each diagram so that the correspond to the
points of the persistence diagrams of the centred image. Then scale the
birth and death times of each point by some factor.
}
