% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendedPersistence.R
\name{extendedPersistence}
\alias{extendedPersistence}
\title{Extended Persistent Homology Transform}
\usage{
extendedPersistence(
  bdryCurves,
  imgName,
  nDirections,
  tolerance = 1/sqrt(2),
  collinearTol = 1e-08,
  saveOutput = FALSE,
  outputDir = NULL,
  fName = NULL,
  verbose = TRUE
)
}
\arguments{
\item{bdryCurves}{A list containing points along the boundary curves of the
image. The list must be structured as
\itemize{
\item \code{bdryCurves[[1]]} is a vector \eqn{(m,n)} where \eqn{m} is the number of
anticlockwise curves and \eqn{n} is the number of clockwise curves.
\item \code{bdryCurves[[i]]} is a matrix containing the points on an anticlockwise
curve for \eqn{2\leq i\leq m+1}.
\item \code{bdryCurves[[j]]} is a matrix containing the points on a clockwise curve
for \eqn{m+2\leq j\leq n+1}.
The length of the list must be \eqn{m+n+1}.
}}

\item{imgName}{A `string`` giving the name of the image.}

\item{nDirections}{An \code{int} giving the number of directions to use. This
must be \emph{positive and even}.}

\item{tolerance}{\code{numeric} parameter to control the noise from the image.
If the distance between two adjacent vertices is less than this value, then
any class generated by them will be ignored. The default value is
\eqn{1/\sqrt{2}}.}

\item{collinearTol}{\code{numeric} parameter to control the tolerance when
determining if three points are collinear. This is used when determining
whether a vertex on a curve is a local minimum. The default value is \code{1e-8}.}

\item{saveOutput}{If TRUE, will save output as a \code{.RDS} file to
directory specified by outputDir. The default values is \code{FALSE}.}

\item{outputDir}{The directory to save the output. If saveOutput is TRUE
and no directory is specified, saves to working directory. The default
is \code{NULL}.}

\item{fName}{The name of the output file saved. If saveOutput is TRUE and
no filename specified, prompts user for filename. The default value is
\code{NULL}.}

\item{verbose}{If TRUE, prints indictors of progress throughout. The default
value is \code{TRUE}.}
}
\value{
A list of length \code{nDirections} where the \emph{i}-th entry is the extended
persistence diagram in the \emph{i}-th direction. Each diagram is a list of
length five with names \emph{"Name"}, \emph{"Ord0"}, \emph{"Rel1"}, \emph{"Ext0"},
\emph{"Ext1"}. The entry for \code{Name} is a string identifying the object and the
direction used for the height filtration. The remaining entries are
(possibly empty) matrices containing the points on the relevant diagram.
}
\description{
The function \code{\link[=extendedPersistence]{extendedPersistence()}} computes the extended persistent
extended persistent homology transform of the boundary curves of a two
dimensional shape in a given number of directions.
}
\details{
Given a disjoint collection of simple closed curves representing the
boundaries of a two dimensional shape, compute the extended
persistent homology transform of the shape in a given number of directions.
We assume the points on each boundary curves are ordered and construct the
one skeleton by adding an edge between each pair of consecutive vertices.
This skeleton is used to compute the height filtration in a given direction
and determine the lower neighbours of each vertex. The regular persistent
homology of each curve is then computed by the union find algorithm and used
to contruct the extended persistent homology.
}
