% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendedPersistence.R
\name{multiExtendedPersistence}
\alias{multiExtendedPersistence}
\title{Compute the Extended PHT for multiple images}
\usage{
multiExtendedPersistence(
  inputDir,
  nDirections,
  tolerance = 1/sqrt(2),
  collinearTol = 1e-08,
  saveOutput = FALSE,
  outputDir = NULL,
  verbose = TRUE
)
}
\arguments{
\item{inputDir}{The directory containing the images.}

\item{nDirections}{An \code{int} giving the number of directions to use. This
must be \emph{positive and even}.}

\item{tolerance}{\code{numeric} parameter to control the noise from the image.
If the distance between two adjacent vertices is less than this value, then
any class generated by them will be ignored. The default value is
\eqn{1/\sqrt{2}}.}

\item{collinearTol}{\code{numeric} parameter to control the tolerance when
determining if three points are collinear. This is used when determining
whether a vertex on a curve is a local minimum. The default value is \code{1e-8}.}

\item{saveOutput}{If TRUE, will save output as a \code{.RDS} file to
directory specified by outputDir. The default values is \code{FALSE}.}

\item{outputDir}{The directory to save the output. If saveOutput is TRUE
and no directory is specified, saves to working directory. The default
is \code{NULL}.}

\item{verbose}{If TRUE, prints indicators of progress throughout. The default
value is \code{TRUE}.}
}
\value{
If \code{saveOutput = TRUE}, then the output of \code{\link[=extendedPersistence]{extendedPersistence()}}
for each object will be saved in a separate file. If \code{saveOutput = FALSE},
then multiExtendedPersistence()] returns a list containing the extended
persistence diagrams for each object in \code{inputDir}. The \eqn{i}-th entry in
the list is the extended persistence diagrams for the \eqn{i}-th file in
\code{inputDir}.
}
\description{
Runs \code{\link[=extendedPersistence]{extendedPersistence()}} on all .RDS files in a given directory.
}
\seealso{
extendedPersistence()
}
